chore: move useErrorMsg file path
